Addresses comments (more in desc)
Put test in correct location

Make new code run on all paths instead of just stride contains a 0

Improve code to work with TensorWrapper
